// src/types/openapi/models/EventHistoryDto.ts
// tslint:disable
/**
 * Core API
 * REST API for CZERTAINLY Core
 *
 * The version of the OpenAPI document: 2.17.1-SNAPSHOT
 * Contact: info@otilm.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { EventStatus, PaginationResponseDtoTriggerHistoryObjectSummaryDto, Resource } from './';

/**
 * @export
 * @interface EventHistoryDto
 */
export interface EventHistoryDto {
    /**
     * Time at which the event firing started.
     * @type {string}
     * @memberof EventHistoryDto
     */
    startedAt: string;
    /**
     * Time at which the event firing finished.
     * @type {string}
     * @memberof EventHistoryDto
     */
    finishedAt?: string;
    /**
     * Status of the event firing.
     * @type {EventStatus}
     * @memberof EventHistoryDto
     */
    status: EventStatus;
    /**
     * Resource of the objects that triggers have been evaluated on.
     * @type {Resource}
     * @memberof EventHistoryDto
     */
    resource: Resource;
    /**
     * Number of objects evaluated.
     * @type {number}
     * @memberof EventHistoryDto
     */
    objectsEvaluated: number;
    /**
     * Number of objects matched at least by one trigger.
     * @type {number}
     * @memberof EventHistoryDto
     */
    objectsMatched: number;
    /**
     * Number of objects matched by ignore triggers.
     * @type {number}
     * @memberof EventHistoryDto
     */
    objectsIgnored: number;
    /**
     * Paginated list of trigger evaluations on individual objects.
     * @type {PaginationResponseDtoTriggerHistoryObjectSummaryDto}
     * @memberof EventHistoryDto
     */
    objectHistories: PaginationResponseDtoTriggerHistoryObjectSummaryDto;
}